Implemented GetRandom invalid_argument exception and placed it into its own header file.
