+++ Matchmaking Player Generator/Matchmaking Player Generator.vcxproj.filters
@@ -26,5 +26,8 @@
     <ClInclude Include="Player.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="GetRandom.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
